@@ -3,7 +3,7 @@
   "description": "Material UI based dynamic form component for React using JSON-Schema",
   "author": "Dinakaran S",
   "user": "dinakarans",
-  "version": "1.7.0-beta",
+  "version": "1.7.4-beta",
   "scripts": {
     "start": "webpack-dev-server --port 8000",
     "test": "jest --",
@@ -25,15 +25,15 @@
   "module": "src",
   "jsnext:main": "src",
   "devDependencies": {
-    "@babel/core": "^7.8.4",
+    "@babel/core": "^7.8.7",
     "@babel/plugin-syntax-dynamic-import": "^7.8.3",
     "@babel/plugin-transform-modules-commonjs": "^7.8.3",
-    "@babel/register": "^7.8.3",
-    "@babel/standalone": "^7.8.4",
+    "@babel/register": "^7.8.6",
+    "@babel/standalone": "^7.8.7",
     "@catalog/markdown-loader": "0.0.5",
     "@date-io/date-fns": "^1.3.11",
     "babel-core": "^7.0.0-bridge.0",
-    "babel-eslint": "^10.0.3",
+    "babel-eslint": "^10.1.0",
     "babel-jest": "^25.1.0",
     "babel-loader": "^8.0.6",
     "babel-plugin-transform-react-remove-prop-types": "^0.4.24",
@@ -42,14 +42,14 @@
     "chai": "^4.2.0",
     "clean-webpack-plugin": "^3.0.0",
     "css-loader": "^3.4.2",
-    "date-fns": "^2.9.0",
+    "date-fns": "^2.10.0",
     "eslint": "^6.8.0",
     "eslint-config-airbnb": "^18.0.1",
     "eslint-loader": "^3.0.3",
     "eslint-plugin-import": "^2.20.1",
     "eslint-plugin-jsx-a11y": "^6.2.3",
-    "eslint-plugin-react": "^7.18.3",
-    "file-loader": "^5.0.2",
+    "eslint-plugin-react": "^7.19.0",
+    "file-loader": "^5.1.0",
     "gh-pages": "^2.2.0",
     "git-prepush-hook": "^1.0.2",
     "html-webpack-plugin": "^3.2.0",
@@ -58,31 +58,31 @@
     "jquery": "^3.4.1",
     "json-loader": "^0.5.7",
     "mini-css-extract-plugin": "^0.9.0",
-    "npm": "^6.13.7",
+    "npm": "^6.14.2",
     "optimize-css-assets-webpack-plugin": "^5.0.3",
     "prop-types": "^15.7.2",
     "purecss": "^1.0.1",
     "raw-loader": "^4.0.0",
     "react": "^16.8.6",
     "react-addons-test-utils": "^15.6.2",
-    "react-dom": "^16.12.0",
+    "react-dom": "^16.13.0",
     "react-github-corner": "^2.3.0",
     "react-highlight": "^0.12.0",
     "react-json-tree": "^0.11.2",
     "react-split-pane": "^0.1.89",
     "regenerator-runtime": "^0.13.3",
-    "rimraf": "^3.0.1",
+    "rimraf": "^3.0.2",
     "style-loader": "^1.1.3",
     "sync-exec": "^0.6.2",
-    "terser-webpack-plugin": "^2.3.4",
+    "terser-webpack-plugin": "^2.3.5",
     "url-loader": "^3.0.0",
-    "webpack": "^4.41.5",
-    "webpack-cli": "^3.3.10",
+    "webpack": "^4.42.0",
+    "webpack-cli": "^3.3.11",
     "webpack-dev-server": "^3.10.3",
     "webpack-merge": "^4.2.2"
   },
   "peerDependencies": {
-    "react": ">=16.12.0"
+    "react": ">=16.13.0"
   },
   "repository": {
     "type": "git",
@@ -120,9 +120,10 @@
   ],
   "dependencies": {
     "@babel/cli": "^7.8.4",
-    "@babel/preset-env": "^7.8.4",
+    "@babel/core": "^7.0.0",
+    "@babel/preset-env": "^7.8.7",
     "@babel/preset-react": "^7.8.3",
-    "@material-ui/core": "^4.9.1",
+    "@material-ui/core": "^4.9.5",
     "@material-ui/pickers": "^3.2.10",
     "autosuggest-highlight": "^3.1.1",
     "classnames": "^2.2.6",
