@@ -1,11 +1,11 @@
-/* eslint-disable */
 // adapted based on rackt/history (MIT)
 // Node 0.10+
-var execSync = require('child_process').execSync;
-var stat = require('fs').stat;
+let execSync = require('child_process').execSync;
+const stat = require('fs').stat;
 
 // Node 0.10 check
 if (!execSync) {
+  // eslint-disable-next-line global-require
   execSync = require('sync-exec');
 }
 
@@ -15,14 +15,15 @@ function exec(command) {
   });
 }
 
-stat('dist-modules', function(error, stat) {
+// eslint-disable-next-line no-shadow
+stat('dist-modules', function (error, stat) {
   // Skip building on Travis
   if (process.env.TRAVIS) {
     return;
   }
 
   if (error || !stat.isDirectory()) {
-    exec('npm i @babel/cli @babel/core @babel/preset-env @babel/preset-react');
+    exec('npm i @babel/cli @babel/core @babel/preset-env @babel/react');
     exec('npm run dist:modules');
   }
 });
